feat: StudyList GET 1차 완성
